an adding machine
We don't even need to touch the second summand;
the price for that is that the results shift on tape,
i.e. the sum is written two bits to the right
of the original position of the first summand.
The advantaga is in sums like 0011101111111111111111111111100.
The disadvantage is in repeated sums shifting.